Instead of blocking animated gifs, just stop them
This will just stop animated gifs from animating but there is
no way to figure out if the image I am stopping is a gif. If
there was a way, I could have added a play element on top of it
and it will allow a user to animate it again.